modularized models.py / removed models.py
