+++ minance/api/views.py
@@ -12,7 +12,7 @@ from flask import (
 )
 
 from minance import create_app
-from minance.models import Asset, Candle
+from minance.asset.models import Asset, Candle
 
 api_old = Blueprint("api_old", __name__)
 api = api_old
