+++ minance/auth/views.py
@@ -2,7 +2,7 @@ from flask import Blueprint, render_template, redirect, url_for, request, flash
 from flask_login import login_user, logout_user, current_user
 
 from minance import db, bcrypt
-from minance.models import User
+from minance.user.models import User
 
 auth = Blueprint("auth", __name__)
